edit or delete education onclick of tr

+++ client/src/redux/actions/profile.js
@@ -154,13 +154,14 @@ export const addExperience = (formData, history) => async dispatch => {
 };
 
 // Add Education
-export const addEducation = (formData, history) => async dispatch => {
+export const addEducation = (formData, history, id = null) => async dispatch => {
   try {
     const config = {
       headers: {
         'Content-Type': 'application/json'
       }
     };
+    if (id) formData.id = id
 
     const res = await axios.put('/api/profile/education', formData, config);
 
@@ -168,8 +169,7 @@ export const addEducation = (formData, history) => async dispatch => {
       type: UPDATE_PROFILE,
       payload: res.data
     });
-
-    dispatch(setAlert('Education Added', 'success'));
+    dispatch(setAlert(id ? 'Education Updated' : 'Education Added', 'success'));
 
     history.push('/dashboard');
   } catch (err) {
@@ -206,16 +206,21 @@ export const deleteExperience = id => async dispatch => {
 };
 
 // Delete education
-export const deleteEducation = id => async dispatch => {
+export const deleteEducation = (id, history = null) => async dispatch => {
   try {
-    const res = await axios.delete(`/api/profile/education/${id}`);
+    if (!window.confirm('Are you sure you want to delete this?')) {
+      return history.push(`/education/edit/${id}`)
+    }
 
+    const res = await axios.delete(`/api/profile/education/${id}`);
     dispatch({
       type: UPDATE_PROFILE,
       payload: res.data
     });
-
+    
+    history.push('/dashboard')
     dispatch(setAlert('Education Removed', 'success'));
+
   } catch (err) {
     dispatch({
       type: PROFILE_ERROR,
